Reorganizing types.
